# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: BugTracker API
  version: 1.0.0
  description: |
    REST API für den BugTracker — eine Webanwendung zur Verwaltung von Software-Fehlern.

    **Authentifizierung:** Session-Cookie (`JSESSIONID`). Login per `POST /api/auth/login`,
    der Server setzt das Cookie. Folgende Requests müssen das Cookie mitsenden
    (Browser macht das automatisch).

    **CSRF:** Geschützte Endpoints (alles außer Login/Register) erwarten den
    CSRF-Token im Header `X-XSRF-TOKEN`. Der Token wird vom Server als Cookie
    `XSRF-TOKEN` gesetzt — das Frontend liest ihn aus und sendet ihn im Header
    mit. Siehe TASKS.md T021 für CSRF-Konzept.

    **Rollen:**
    - `TESTER` — kann Bugs anlegen, sehen, suchen, kommentieren
    - `DEVELOPER` — zusätzlich Bugs bearbeiten, Status/Priorität/Bearbeiter ändern
    - `ADMIN` — zusätzlich Tags und User verwalten

    **Status-State-Machine (FA-06, T036):**
    - `NEU → IN_BEARBEITUNG, ABGELEHNT`
    - `IN_BEARBEITUNG → IM_REVIEW, ABGELEHNT`
    - `IM_REVIEW → ERLEDIGT, IN_BEARBEITUNG, ABGELEHNT`
    - `ERLEDIGT → ARCHIVIERT`
    - `ABGELEHNT → ARCHIVIERT`
    - alle Status `→ ARCHIVIERT`

    Verbotene Übergänge geben `400 Bad Request` zurück.
  contact:
    name: BugTracker Team — Gruppe DE 3 · MCI Innsbruck
  license:
    name: Universitätsprojekt — MCI Innsbruck
servers:
  - url: http://localhost:8080
    description: Lokale Entwicklung
  - url: https://bugtracker.example.com
    description: Produktion (geplant, T072)
security:
  - sessionAuth: []
  - csrfToken: []
...
paths:
...
  /api/bugs:
    get:
      tags: [Bugs]
      summary: Bug-Liste mit Filtern (FA-02, FA-09, T032)
      description: |
        Liefert eine paginierte Liste von Bugs.

        Filter sind kombinierbar — alle als Query-Parameter.
        `status` kann mehrfach angegeben werden (z.B. `?status=NEU&status=IN_BEARBEITUNG`).

        **Default:** `archived=false` (archivierte Bugs werden ausgeblendet).

        **Suche:** `search` macht case-insensitive ILIKE auf `title`.
      operationId: listBugs
      parameters:
        - name: status
          in: query
          description: Nach Status filtern (multi)
          schema:
            type: array
            items:
              $ref: '#/components/schemas/BugStatus'
          style: form
          explode: true
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/Priority'
        - name: assignee_id
          in: query
          description: Nach Bearbeiter-ID filtern
          schema:
            type: integer
            format: int64
        - name: tagIds
          in: query
          description: Nach Tag-IDs filtern (multi, OR-Semantik — Bug matched, wenn mind. ein Tag passt)
          schema:
            type: array
            items:
              type: integer
              format: int64
          style: form
          explode: true
        - name: search
          in: query
          description: Volltext-Suche im Titel (FA-10, T060)
          schema:
            type: string
        - name: archived
          in: query
          description: Archivierte einschließen (default false)
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Bug-Liste
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BugListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Bugs]
      summary: Bug anlegen (FA-01, T031)
      description: |
        Legt einen neuen Bug an. Reporter = eingeloggter User.
        Default-Status: `NEU`. Default-Priorität: `MITTEL`.

        **Rolle:** TESTER, DEVELOPER, ADMIN
      operationId: createBug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBugRequest'
            example:
              title: Login-Button reagiert nicht
              description: Beim Klick auf Login passiert nichts. Reproduktion ...
              priority: HOCH
              tagId: 1
      responses:
        '201':
          description: Bug erfolgreich erstellt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bug'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/bugs/{id}:
    parameters:
      - $ref: '#/components/parameters/BugIdPath'

    get:
      tags: [Bugs]
      summary: Bug-Detail (FA-03, T033)
      operationId: getBug
      responses:
        '200':
          description: Bug-Details inkl. Reporter, Assignee, Tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bug'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Bugs]
      summary: Bug bearbeiten (FA-04, T034)
      description: |
        Bearbeitet Titel, Beschreibung und Tag-Zuordnungen.
        Priorität ändert sich via separatem Endpoint (PATCH /priority).
        Activity-Einträge folgen mit FA-14 (Migration T056).

        **Rolle:** DEVELOPER, ADMIN, sowie TESTER für eigene Bugs (Reporter).
      operationId: updateBug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBugRequest'
      responses:
        '200':
          description: Bug aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bug'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Bug ist archiviert und kann nicht bearbeitet werden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/bugs/{id}/archive:
    parameters:
      - $ref: '#/components/parameters/BugIdPath'
    patch:
      tags: [Bugs]
      summary: Bug archivieren — Soft-Delete (FA-05, T035)
      description: |
        Setzt `archived = true` und `status = ARCHIVIERT`. Der Bug bleibt in der DB
        erhalten und kann via PATCH /restore wieder reaktiviert werden. Erzeugt
        einen Activity-Eintrag (`field=archived`, `false`→`true`).

        **Rolle:** DEVELOPER, ADMIN
      operationId: archiveBug
      responses:
        '200':
          description: Bug archiviert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bug'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Bug ist bereits archiviert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/bugs/{id}/restore:
    parameters:
      - $ref: '#/components/parameters/BugIdPath'
    patch:
      tags: [Bugs]
      summary: Bug reaktivieren (FA-05, T035)
      description: |
        Setzt `archived = false` und `status = NEU` (Workflow-Reset, da `ARCHIVIERT`
        terminal ist). Erzeugt einen Activity-Eintrag (`field=archived`, `true`→`false`).

        **Rolle:** DEVELOPER, ADMIN
      operationId: restoreBug
      responses:
        '200':
          description: Bug reaktiviert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bug'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Bug ist nicht archiviert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/bugs/{id}/status:
    parameters:
      - $ref: '#/components/parameters/BugIdPath'
    patch:
      tags: [Bugs]
      summary: Status ändern (FA-06, T036)
      description: |
        Ändert den Status mit State-Machine-Validierung. Verbotene Übergänge
        geben `400` zurück.

        Erlaubte Übergänge — siehe API-Beschreibung oben.

        **Rolle:** DEVELOPER, ADMIN
      operationId: updateBugStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
            example:
              status: IN_BEARBEITUNG
      responses:
        '200':
          description: Status aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bug'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: |
            Statuswechsel nicht erlaubt (State-Machine-Verletzung) ODER
            Bug ist archiviert. Im Body steht welcher.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                stateMachine:
                  value:
                    error: "Status-Wechsel nicht erlaubt: ERLEDIGT → NEU"
                archived:
                  value:
                    error: "Archivierter Bug kann nicht bearbeitet werden"
...
  /api/bugs/{id}/assignee:
    parameters:
      - $ref: '#/components/parameters/BugIdPath'
    patch:
      tags: [Bugs]
      summary: Bearbeiter zuweisen (FA-08, T037)
      description: |
        Setzt oder entfernt den Bearbeiter. `assigneeId: null` entfernt den Bearbeiter.

        **Rolle:** DEVELOPER, ADMIN
      operationId: updateBugAssignee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssigneeRequest'
            example:
              assigneeId: 3
      responses:
        '200':
          description: Bearbeiter aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bug'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Bug oder User nicht gefunden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============================================================
  # ACTIVITIES
  # ============================================================
...
components:
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Spring Boot Session-Cookie. Wird beim Login gesetzt.
    csrfToken:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: |
        CSRF-Token aus dem `XSRF-TOKEN`-Cookie. Pflicht für alle nicht-GET-Requests
        außer Login/Register/Logout. Siehe TASKS.md T021.

  # ----------------------------------------------------------
  # PARAMETERS
  # ----------------------------------------------------------
  parameters:
    BugIdPath:
      name: id
      in: path
      required: true
      description: Bug-ID
      schema:
        type: integer
        format: int64
        example: 42

  # ----------------------------------------------------------
  # RESPONSES (wiederverwendbar)
  # ----------------------------------------------------------
  responses:
    Unauthorized:
      description: Nicht eingeloggt oder Session abgelaufen
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Nicht authentifiziert
    Forbidden:
      description: Keine Berechtigung für diese Aktion
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Keine Berechtigung
    NotFound:
      description: Ressource nicht gefunden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Nicht gefunden
    ValidationError:
      description: Validierungsfehler im Request-Body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            error: Validierung fehlgeschlagen
            fieldErrors:
              title: Titel ist erforderlich
              priority: Ungültige Priorität

  # ----------------------------------------------------------
  # SCHEMAS — Enums
  # ----------------------------------------------------------
  schemas:
    BugStatus:
      type: string
      description: |
        Status eines Bugs. Siehe State-Machine-Diagramm in API-Beschreibung.
      enum:
        - NEU
        - IN_BEARBEITUNG
        - IM_REVIEW
        - ERLEDIGT
        - ABGELEHNT
        - ARCHIVIERT
    Priority:
      type: string
      description: Priorität eines Bugs (FA-07).
      enum:
        - NIEDRIG
        - MITTEL
        - HOCH
        - KRITISCH
...
    Bug:
      type: object
      required:
        - id
        - title
        - status
        - priority
        - reporterId
        - reporterName
        - archived
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          example: 42
        title:
          type: string
          maxLength: 255
          example: Login-Button reagiert nicht
        description:
          type: string
          example: Beim Klick auf Login passiert nichts.
        status:
          $ref: '#/components/schemas/BugStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        reporterId:
          type: integer
          format: int64
          example: 1
        reporterName:
          type: string
          description: Username des Reporters (Convenience-Feld für Frontend)
          example: marie
        assigneeId:
          type: integer
          format: int64
          nullable: true
          example: 3
        assigneeName:
          type: string
          nullable: true
          example: tom
        tagId:
          type: integer
          format: int64
          nullable: true
          example: 1
        tagName:
          type: string
          nullable: true
          example: Backend
        archived:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateBugRequest:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          minLength: 1
        priority:
          $ref: '#/components/schemas/Priority'
        tagId:
          type: integer
          format: int64
          nullable: true
    UpdateBugRequest:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          minLength: 1
          maxLength: 10000
        tagIds:
          type: array
          items:
            type: integer
            format: int64
          description: |
            M:N — komplette Tag-Zuordnung. Wenn `null`, bleiben bestehende Tags erhalten;
            leere Liste entfernt alle Tags.
          nullable: true
    UpdateStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/BugStatus'
...
    UpdateAssigneeRequest:
      type: object
      required: [assigneeId]
      properties:
        assigneeId:
          type: integer
          format: int64
          nullable: true
          description: User-ID oder `null` zum Entfernen
...
    BugListResponse:
      type: object
      required: [bugs, total, page, pageSize]
      properties:
        bugs:
          type: array
          items:
            $ref: '#/components/schemas/Bug'
        total:
          type: integer
          example: 73
        page:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 50

    # --------------------------------------------------------
    # SCHEMAS — Errors
    # --------------------------------------------------------
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Menschen-lesbare Fehlermeldung
          example: Nicht gefunden
    ValidationError:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Validierung fehlgeschlagen
        fieldErrors:
          type: object
          description: Map von Feldname → Fehlermeldung
          additionalProperties:
            type: string
          example:
            title: Titel ist erforderlich
            email: Ungültiges E-Mail-Format